feat(metrics): refactor metrics imports
